ref: move pagination methods to listItem/pagination
